Memory allocator working on 32-bit

+++ lib_src/string.c
@@ -1,3 +1,5 @@
+#include "stddef.h"
+
 int strlen(char s[]) {
     int len = 0;
     while (s[len]) ++len;
@@ -23,22 +25,22 @@ int strcmp(char s1[], char s2[]) {
     return s1[i] - s2[i];
 }
 
-int memcmp(const void *str1, const void *str2, unsigned int n) {
+int memcmp(const void *str1, const void *str2, size_t n) {
     char *s1, *s2;
     s1 = str1;
     s2 = str2;
-    for (unsigned int i = 0; s1[i] == s2[i]; ++i) {
+    for (size_t i = 0; s1[i] == s2[i]; ++i) {
         if ((i + 1) == n) return 0;
     }
     return s1[i] - s2[i];
 }
 
-void *memset(char *s, char c, unsigned int n) {
+void *memset(char *s, char c, size_t n) {
     for (unsigned int i = 0; i < n; ++i)
         s[i] = c;
 }
 
-void *memcpy(void *s, const void *ct, unsigned int n) {
+void *memcpy(void *s, const void *ct, size_t n) {
     do {
         --n;
         ((char *) s)[n] = ((char *) ct)[n];
